Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -13,7 +13,7 @@ int main() {
      for(int i=0;i<n;i++){
         sum+=c[i];
      }
-     int ans=sum/n;
+     float ans=sum/n;
         printf("Average Marks: %.2f",ans);
     return 0;
 }
